Face Mask Detection/requirements.txt
# Bibliotecas principais
numpy>=1.19.5
pandas>=1.2.4
matplotlib>=3.4.2
seaborn>=0.11.1

# Visão Computacional
opencv-python>=4.5.3.56
dlib>=19.22.0
face-recognition>=1.3.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.5.0

# Processamento de Imagens
Pillow>=8.2.0
scikit-image>=0.18.1

# Machine Learning
scikit-learn>=0.24.2

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=5.5.5

# Utilitários
tqdm>=4.61.0
